requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.46
granian==2.7.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2026.1.post1
redis==7.3.0
referencing==0.37.0
reflex==0.8.27
reflex-hosting-cli==0.1.62
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
watchfiles==1.1.1
wrapt==2.1.2
wsproto==1.3.2
